refactor(model builders) add Person base class add parameterized constructors to OrderBuilders change drivingLicense param type to long (DriverBuilder)

+++ CabWebApi/Content/Builders/DriverBuilder.cs
@@ -3,16 +3,16 @@
 namespace CabWebApi.Content.Builders;
 public class DriverBuilder : UserBuilder<Driver, DriverBuilder>
 {
-    public DriverBuilder() : base() { }
-    public DriverBuilder(Driver driver) : base(driver) { }
-    public DriverBuilder Earns(int salary)
-    {
-        user.Salary = salary;
-        return this;
-    }
-    public DriverBuilder HasLicense(int drivingLicense)
-    {
-        user.DrivingLicense = drivingLicense;
-        return this;
-    }
+	public DriverBuilder() : base() { }
+	public DriverBuilder(Driver driver) : base(driver) { }
+	public DriverBuilder Earns(int salary)
+	{
+		user.Salary = salary;
+		return this;
+	}
+	public DriverBuilder HasLicense(long drivingLicense)
+	{
+		user.DrivingLicense = drivingLicense;
+		return this;
+	}
 }
